+++ network-store-iidm-impl/src/test/java/com/powsybl/network/store/iidm/impl/extensions/ActivePowerControlExtensionTest.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.network.store.iidm.impl;
+package com.powsybl.network.store.iidm.impl.extensions;
 
 import com.powsybl.commons.extensions.Extension;
 import com.powsybl.iidm.network.Battery;
@@ -13,6 +13,8 @@ import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.NetworkListener;
 import com.powsybl.iidm.network.extensions.ActivePowerControl;
 import com.powsybl.iidm.network.extensions.ActivePowerControlAdder;
+import com.powsybl.network.store.iidm.impl.BatteryImpl;
+import com.powsybl.network.store.iidm.impl.CreateNetworksUtil;
 import org.junit.Test;
 
 import java.util.List;
